+++ java/src/test/java/com/nubiform/java/stream/DishTest.java
@@ -3,9 +3,7 @@ package com.nubiform.java.stream;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import java.util.Arrays;
 import java.util.List;
-import java.util.stream.Stream;
 
 import static java.util.stream.Collectors.toList;
 import static org.junit.jupiter.api.Assertions.*;
@@ -13,10 +11,6 @@ import static org.junit.jupiter.api.Assertions.*;
 class DishTest {
 
     List<Dish> menu;
-
-    List<String> words;
-
-    String[] arrayOfWorld;
 
     @BeforeEach
     public void setUp() {
@@ -31,15 +25,6 @@ class DishTest {
                 new Dish("prawns", false, 300, Dish.Type.FISH),
                 new Dish("salmon", false, 450, Dish.Type.FISH)
         );
-
-        words = List.of("Hello", "World");
-
-        arrayOfWorld = new String[] {"Goodbye", "World"};
-    }
-
-    @Test
-    public void test() {
-        assertTrue(true);
     }
 
     @Test
@@ -57,28 +42,6 @@ class DishTest {
     public void vegetarianDishes() {
         List<Dish> collect = menu.stream()
                 .filter(Dish::isVegetarian)
-                .collect(toList());
-
-        collect.forEach(System.out::println);
-    }
-
-    @Test
-    public void words() {
-        List<String[]> collect1 = words.stream()
-                .map(word -> word.split(""))
-                .distinct()
-                .collect(toList());
-
-        List<Stream<String>> collect2 = words.stream()
-                .map(word -> word.split(""))
-                .map(Arrays::stream)
-                .distinct()
-                .collect(toList());
-
-        List<String> collect = words.stream()
-                .map(word -> word.split(""))
-                .flatMap(Arrays::stream)
-                .distinct()
                 .collect(toList());
 
         collect.forEach(System.out::println);
